requirements/requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --no-strip-extras --constraint=requirements/common-constraints.txt --output-file=requirements/requirements-dev.txt requirements/requirements-dev.in
build==1.3.0
    # via
    #   -c requirements/common-constraints.txt
    #   pip-tools
cfgv==3.5.0
    # via
    #   -c requirements/common-constraints.txt
    #   pre-commit
click==8.3.1
    # via
    #   -c requirements/common-constraints.txt
    #   pip-tools
    #   typer
codespell==2.4.1
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
cogapp==3.6.0
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
contourpy==1.3.3
    # via
    #   -c requirements/common-constraints.txt
    #   matplotlib
cycler==0.12.1
    # via
    #   -c requirements/common-constraints.txt
    #   matplotlib
distlib==0.4.0
    # via
    #   -c requirements/common-constraints.txt
    #   virtualenv
filelock==3.20.0
    # via
    #   -c requirements/common-constraints.txt
    #   virtualenv
fonttools==4.60.1
    # via
    #   -c requirements/common-constraints.txt
    #   matplotlib
identify==2.6.15
    # via
    #   -c requirements/common-constraints.txt
    #   pre-commit
imgcat==0.6.0
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
iniconfig==2.3.0
    # via
    #   -c requirements/common-constraints.txt
    #   pytest
jinja2==3.1.6
    # via
    #   -c requirements/common-constraints.txt
    #   memray
kiwisolver==1.4.9
    # via
    #   -c requirements/common-constraints.txt
    #   matplotlib
linkify-it-py==2.0.3
    # via
    #   -c requirements/common-constraints.txt
    #   markdown-it-py
lox==1.0.0
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
markdown-it-py[linkify]==4.0.0
    # via
    #   -c requirements/common-constraints.txt
    #   mdit-py-plugins
    #   rich
    #   textual
markupsafe==3.0.3
    # via
    #   -c requirements/common-constraints.txt
    #   jinja2
matplotlib==3.10.7
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
mdit-py-plugins==0.5.0
    # via
    #   -c requirements/common-constraints.txt
    #   textual
mdurl==0.1.2
    # via
    #   -c requirements/common-constraints.txt
    #   markdown-it-py
memray==1.19.2
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
nodeenv==1.9.1
    # via
    #   -c requirements/common-constraints.txt
    #   pre-commit
numpy==2.3.5
    # via
    #   -c requirements/common-constraints.txt
    #   contourpy
    #   matplotlib
    #   pandas
objgraph==3.6.2
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
packaging==25.0
    # via
    #   -c requirements/common-constraints.txt
    #   build
    #   matplotlib
    #   pytest
pandas==2.3.3
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
pillow==12.0.0
    # via
    #   -c requirements/common-constraints.txt
    #   matplotlib
pip==25.3
    # via
    #   -c requirements/common-constraints.txt
    #   pip-tools
pip-tools==7.5.2
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
platformdirs==4.5.0
    # via
    #   -c requirements/common-constraints.txt
    #   textual
    #   virtualenv
pluggy==1.6.0
    # via
    #   -c requirements/common-constraints.txt
    #   pytest
pre-commit==4.5.0
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
pygments==2.19.2
    # via
    #   -c requirements/common-constraints.txt
    #   pytest
    #   rich
    #   textual
pyparsing==3.2.5
    # via
    #   -c requirements/common-constraints.txt
    #   matplotlib
pyproject-hooks==1.2.0
    # via
    #   -c requirements/common-constraints.txt
    #   build
    #   pip-tools
pytest==9.0.1
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
    #   pytest-asyncio
    #   pytest-env
    #   pytest-mock
pytest-asyncio==1.3.0
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
pytest-env==1.2.0
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
pytest-mock==3.15.1
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
python-dateutil==2.9.0.post0
    # via
    #   -c requirements/common-constraints.txt
    #   matplotlib
    #   pandas
pytz==2025.2
    # via
    #   -c requirements/common-constraints.txt
    #   pandas
pyyaml==6.0.3
    # via
    #   -c requirements/common-constraints.txt
    #   pre-commit
rich==14.2.0
    # via
    #   -c requirements/common-constraints.txt
    #   memray
    #   textual
    #   typer
semver==3.0.4
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
setuptools==80.9.0
    # via
    #   -c requirements/common-constraints.txt
    #   pip-tools
shellingham==1.5.4
    # via
    #   -c requirements/common-constraints.txt
    #   typer
six==1.17.0
    # via
    #   -c requirements/common-constraints.txt
    #   python-dateutil
textual==6.8.0
    # via
    #   -c requirements/common-constraints.txt
    #   memray
typer==0.20.0
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
typing-extensions==4.15.0
    # via
    #   -c requirements/common-constraints.txt
    #   pytest-asyncio
    #   textual
    #   typer
tzdata==2025.2
    # via
    #   -c requirements/common-constraints.txt
    #   pandas
uc-micro-py==1.0.3
    # via
    #   -c requirements/common-constraints.txt
    #   linkify-it-py
uv==0.9.11
    # via
    #   -c requirements/common-constraints.txt
    #   -r requirements/requirements-dev.in
virtualenv==20.35.4
    # via
    #   -c requirements/common-constraints.txt
    #   pre-commit
wheel==0.45.1
    # via
    #   -c requirements/common-constraints.txt
    #   pip-tools
